feature: split out modules into standalone packages
The Bio.IO package (under `bio_ex_io`) was a little more trouble to
split out than the others. That's namely because I wanted to integrate
it more tightly with the Bio.Sequence modules originally. However, I
think that the loose coupling is easily achieved by referencing the
behaviours of the `bio_ex_sequences` package.

I've added that module as a test dependency to ensure that we have
integration tests run. I'll want to look into running all integration
tests for the packages on release.